Move parseSchema, parseColumns, parseSchemaGroups & distributeFieldsInGroups in a mixin
https://trello.com/c/0FCmhb9F/31-move-parseschema-and-parsecolumns-code-in-a-mixin
